# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Do Quintal à Cozinha API",
    "version": "1.0.0",
    "description": "API for managing rural women producers' products, profiles, and media. This platform connects producers from MMTR-SE with consumers.",
    "contact": {
      "name": "MMTR-SE",
      "url": "https://github.com/catuaba07/do-quintal-a-cozinha"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server (Docker)"
    },
    {
      "url": "https://your-production-domain.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/email": {
      "get": {
        "tags": ["Integration"],
        "summary": "Email contact redirect",
        "description": "Redirects the user to their default email client with a pre-filled recipient and optional subject line. This endpoint performs a 302 redirect to a mailto: URL. The recipient email is configured in the site configuration and never exposed in the frontend.",
        "operationId": "emailRedirect",
        "parameters": [
          {
            "name": "subject",
            "in": "query",
            "description": "Subject line for the email (optional, defaults to 'Contato via site')",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Dúvida sobre produto"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to email client",
            "headers": {
              "Location": {
                "description": "Mailto URL with pre-filled subject",
                "schema": {
                  "type": "string",
                  "format": "uri",
                  "example": "mailto:contato@example.com?subject=D%C3%BAvida%20sobre%20produto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "API_KEY",
        "description": "API key for authentication. Contact the development team to obtain your key."
      }
    },
...
  }
}